requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
cachetools>=5.3.0
psycopg2-binary>=2.9.5